OBG 184 move tests to impl
