+++ fintech-examples/fintech-impl/src/main/java/de/adorsys/opba/fintech/impl/config/FinTechImplConfig.java
@@ -1,15 +1,15 @@
 package de.adorsys.opba.fintech.impl.config;
 
-import de.adorsys.opba.fintech.impl.properties.CookieConfigProperties;
 import de.adorsys.opba.tpp.bankserach.api.resource.generated.TppBankSearchApi;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 @ComponentScan("de.adorsys.opba.fintech.impl")
+@EntityScan(basePackageClasses = {TempEntity.class})
 public class FinTechImplConfig {
 
     @Value("${tpp.url}")
