+++ fintech-examples/fintech-impl/src/main/java/de/adorsys/opba/fintech/impl/database/repositories/UserRepository.java
@@ -1,4 +1,4 @@
-package de.adorsys.opba.fintech.server;
+package de.adorsys.opba.fintech.impl.database.repositories;
 
 import de.adorsys.opba.fintech.impl.service.entities.TempEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
