+++ fintech-examples/fintech-impl/src/test/resources/application.yml
@@ -1,0 +1,6 @@
+spring:
+  datasource:
+    url: jdbc:h2:mem:test
+    username: SA
+    password:
+    driver-class-name: org.h2.Driver
